+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceConfigurationWindowsTeamPolicyWindows10/MSFT_IntuneDeviceConfigurationWindowsTeamPolicyWindows10.psm1
@@ -5,7 +5,108 @@ function Get-TargetResource
     param
     (
         #region resource generator code
-<ParameterBlock><AssignmentsParam>        #endregion
+        [Parameter()]
+        [System.Boolean]
+        $AzureOperationalInsightsBlockTelemetry,
+
+        [Parameter()]
+        [System.String]
+        $AzureOperationalInsightsWorkspaceId,
+
+        [Parameter()]
+        [System.String]
+        $AzureOperationalInsightsWorkspaceKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $ConnectAppBlockAutoLaunch,
+
+        [Parameter()]
+        [System.Boolean]
+        $MaintenanceWindowBlocked,
+
+        [Parameter()]
+        [System.Int32]
+        $MaintenanceWindowDurationInHours,
+
+        [Parameter()]
+        [System.TimeSpan]
+        $MaintenanceWindowStartTime,
+
+        [Parameter()]
+        [System.Boolean]
+        $MiracastBlocked,
+
+        [Parameter()]
+        [ValidateSet('userDefined','one','two','three','four','five','six','seven','eight','nine','ten','eleven','thirtySix','forty','fortyFour','fortyEight','oneHundredFortyNine','oneHundredFiftyThree','oneHundredFiftySeven','oneHundredSixtyOne','oneHundredSixtyFive')]
+        [System.String]
+        $MiracastChannel,
+
+        [Parameter()]
+        [System.Boolean]
+        $MiracastRequirePin,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockMyMeetingsAndFiles,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockSessionResume,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockSigninSuggestions,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsDefaultVolume,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsScreenTimeoutInMinutes,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsSessionTimeoutInMinutes,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsSleepTimeoutInMinutes,
+
+        [Parameter()]
+        [System.String]
+        $WelcomeScreenBackgroundImageUrl,
+
+        [Parameter()]
+        [System.Boolean]
+        $WelcomeScreenBlockAutomaticWakeUp,
+
+        [Parameter()]
+        [ValidateSet('userDefined','showOrganizerAndTimeOnly','showOrganizerAndTimeAndSubject')]
+        [System.String]
+        $WelcomeScreenMeetingInformation,
+
+        [Parameter()]
+        [System.String]
+        $Description,
+
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $DisplayName,
+
+        [Parameter()]
+        [System.Boolean]
+        $SupportsScopeTags,
+
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $Id,
+
+        [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance[]]
+        $Assignments,
+        #endregion
 
         [Parameter()]
         [System.String]
@@ -39,9 +140,9 @@ function Get-TargetResource
 
     try
     {
-        $ConnectionMode = New-M365DSCConnection -Workload '<#Workload#>' `
+        $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
             -InboundParameters $PSBoundParameters `
-            -ProfileName '<#APIVersion#>'
+            -ProfileName 'beta'
 
         #Ensure the proper dependencies are installed in the current environment.
         Confirm-M365DSCDependencies
@@ -58,34 +159,105 @@ function Get-TargetResource
         $nullResult = $PSBoundParameters
         $nullResult.Ensure = 'Absent'
 
-        $getValue = $null<#ResourceGenerator
+        $getValue = $null
         #region resource generator code
-        $getValue = <GetCmdLetName> <getKeyIdentifier> -ErrorAction SilentlyContinue
+        $getValue = Get-MgDeviceManagementDeviceConfiguration -DeviceConfigurationId $Id  -ErrorAction SilentlyContinue
 
         if ($null -eq $getValue)
         {
-            Write-Verbose -Message "Could not find an <ResourceDescription> with <PrimaryKey> {$<PrimaryKey>}"
+            Write-Verbose -Message "Could not find an Intune Device Configuration Windows Team Policy for Windows10 with Id {$Id}"
 
-            if (-Not [string]::IsNullOrEmpty($<FilterKey>))
+            if (-Not [string]::IsNullOrEmpty($DisplayName))
             {
-                $getValue = <GetCmdLetName> `
-<AlternativeFilter>
+                $getValue = Get-MgDeviceManagementDeviceConfiguration `
+                    -Filter "DisplayName eq '$DisplayName'" `
+                    -ErrorAction SilentlyContinue | Where-Object `
+                    -FilterScript { `
+                        $_.AdditionalProperties.'@odata.type' -eq "#microsoft.graph.windows10TeamGeneralConfiguration" `
+                    }
             }
         }
-        #endregionResourceGenerator#>
+        #endregion
         if ($null -eq $getValue)
         {
-            Write-Verbose -Message "Could not find an <ResourceDescription> with <FilterKey> {$<FilterKey>}"
+            Write-Verbose -Message "Could not find an Intune Device Configuration Windows Team Policy for Windows10 with DisplayName {$DisplayName}"
             return $nullResult
         }
-        $<PrimaryKey> = $getValue.<PrimaryKey>
-        Write-Verbose -Message "An <ResourceDescription> with <PrimaryKey> {$<PrimaryKey>} and <FilterKey> {$<FilterKey>} was found."<#ResourceGenerator
-<ComplexTypeConstructor><EnumTypeConstructor><DateTypeConstructor><TimeTypeConstructor>ResourceGenerator#>
-        $results = @{<#ResourceGenerator
-            #region resource generator code
-<HashTableMapping>            #endregionResourceGenerator#>
+        $Id = $getValue.Id
+        Write-Verbose -Message "An Intune Device Configuration Windows Team Policy for Windows10 with Id {$Id} and DisplayName {$DisplayName} was found."
+
+        #region resource generator code
+        $enumMiracastChannel = $null
+        if ($null -ne $getValue.AdditionalProperties.miracastChannel)
+        {
+            $enumMiracastChannel = $getValue.AdditionalProperties.miracastChannel.ToString()
         }
-<#ComplexTypeContent#><#AssignmentsGet#>
+
+        $enumWelcomeScreenMeetingInformation = $null
+        if ($null -ne $getValue.AdditionalProperties.welcomeScreenMeetingInformation)
+        {
+            $enumWelcomeScreenMeetingInformation = $getValue.AdditionalProperties.welcomeScreenMeetingInformation.ToString()
+        }
+        #endregion
+
+        #region resource generator code
+        $timeMaintenanceWindowStartTime = $null
+        if ($null -ne $getValue.AdditionalProperties.maintenanceWindowStartTime)
+        {
+            $timeMaintenanceWindowStartTime = ([TimeSpan]$getValue.AdditionalProperties.maintenanceWindowStartTime).ToString()
+        }
+        #endregion
+
+        $results = @{
+            #region resource generator code
+            AzureOperationalInsightsBlockTelemetry = $getValue.AdditionalProperties.azureOperationalInsightsBlockTelemetry
+            AzureOperationalInsightsWorkspaceId    = $getValue.AdditionalProperties.azureOperationalInsightsWorkspaceId
+            AzureOperationalInsightsWorkspaceKey   = $getValue.AdditionalProperties.azureOperationalInsightsWorkspaceKey
+            ConnectAppBlockAutoLaunch              = $getValue.AdditionalProperties.connectAppBlockAutoLaunch
+            MaintenanceWindowBlocked               = $getValue.AdditionalProperties.maintenanceWindowBlocked
+            MaintenanceWindowDurationInHours       = $getValue.AdditionalProperties.maintenanceWindowDurationInHours
+            MaintenanceWindowStartTime             = $timeMaintenanceWindowStartTime
+            MiracastBlocked                        = $getValue.AdditionalProperties.miracastBlocked
+            MiracastChannel                        = $enumMiracastChannel
+            MiracastRequirePin                     = $getValue.AdditionalProperties.miracastRequirePin
+            SettingsBlockMyMeetingsAndFiles        = $getValue.AdditionalProperties.settingsBlockMyMeetingsAndFiles
+            SettingsBlockSessionResume             = $getValue.AdditionalProperties.settingsBlockSessionResume
+            SettingsBlockSigninSuggestions         = $getValue.AdditionalProperties.settingsBlockSigninSuggestions
+            SettingsDefaultVolume                  = $getValue.AdditionalProperties.settingsDefaultVolume
+            SettingsScreenTimeoutInMinutes         = $getValue.AdditionalProperties.settingsScreenTimeoutInMinutes
+            SettingsSessionTimeoutInMinutes        = $getValue.AdditionalProperties.settingsSessionTimeoutInMinutes
+            SettingsSleepTimeoutInMinutes          = $getValue.AdditionalProperties.settingsSleepTimeoutInMinutes
+            WelcomeScreenBackgroundImageUrl        = $getValue.AdditionalProperties.welcomeScreenBackgroundImageUrl
+            WelcomeScreenBlockAutomaticWakeUp      = $getValue.AdditionalProperties.welcomeScreenBlockAutomaticWakeUp
+            WelcomeScreenMeetingInformation        = $enumWelcomeScreenMeetingInformation
+            Description                            = $getValue.Description
+            DisplayName                            = $getValue.DisplayName
+            SupportsScopeTags                      = $getValue.SupportsScopeTags
+            Id                                     = $getValue.Id
+            Ensure                                 = 'Present'
+            Credential                             = $Credential
+            ApplicationId                          = $ApplicationId
+            TenantId                               = $TenantId
+            ApplicationSecret                      = $ApplicationSecret
+            CertificateThumbprint                  = $CertificateThumbprint
+            Managedidentity                        = $ManagedIdentity.IsPresent
+            #endregion
+        }
+        $assignmentsValues = Get-MgDeviceManagementDeviceConfigurationAssignment -DeviceConfigurationId $Id
+        $assignmentResult = @()
+        foreach ($assignmentEntry in $AssignmentsValues)
+        {
+            $assignmentValue = @{
+                dataType = $assignmentEntry.Target.AdditionalProperties.'@odata.type'
+                deviceAndAppManagementAssignmentFilterType = $(if ($null -ne $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType)
+                    {$assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType.ToString()})
+                deviceAndAppManagementAssignmentFilterId = $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterId
+                groupId = $assignmentEntry.Target.AdditionalProperties.groupId
+            }
+            $assignmentResult += $assignmentValue
+        }
+        $results.Add('Assignments', $assignmentResult)
+
         return [System.Collections.Hashtable] $results
     }
     catch
@@ -106,7 +278,108 @@ function Set-TargetResource
     param
     (
         #region resource generator code
-<ParameterBlock><AssignmentsParam>        #endregion
+        [Parameter()]
+        [System.Boolean]
+        $AzureOperationalInsightsBlockTelemetry,
+
+        [Parameter()]
+        [System.String]
+        $AzureOperationalInsightsWorkspaceId,
+
+        [Parameter()]
+        [System.String]
+        $AzureOperationalInsightsWorkspaceKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $ConnectAppBlockAutoLaunch,
+
+        [Parameter()]
+        [System.Boolean]
+        $MaintenanceWindowBlocked,
+
+        [Parameter()]
+        [System.Int32]
+        $MaintenanceWindowDurationInHours,
+
+        [Parameter()]
+        [System.TimeSpan]
+        $MaintenanceWindowStartTime,
+
+        [Parameter()]
+        [System.Boolean]
+        $MiracastBlocked,
+
+        [Parameter()]
+        [ValidateSet('userDefined','one','two','three','four','five','six','seven','eight','nine','ten','eleven','thirtySix','forty','fortyFour','fortyEight','oneHundredFortyNine','oneHundredFiftyThree','oneHundredFiftySeven','oneHundredSixtyOne','oneHundredSixtyFive')]
+        [System.String]
+        $MiracastChannel,
+
+        [Parameter()]
+        [System.Boolean]
+        $MiracastRequirePin,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockMyMeetingsAndFiles,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockSessionResume,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockSigninSuggestions,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsDefaultVolume,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsScreenTimeoutInMinutes,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsSessionTimeoutInMinutes,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsSleepTimeoutInMinutes,
+
+        [Parameter()]
+        [System.String]
+        $WelcomeScreenBackgroundImageUrl,
+
+        [Parameter()]
+        [System.Boolean]
+        $WelcomeScreenBlockAutomaticWakeUp,
+
+        [Parameter()]
+        [ValidateSet('userDefined','showOrganizerAndTimeOnly','showOrganizerAndTimeAndSubject')]
+        [System.String]
+        $WelcomeScreenMeetingInformation,
+
+        [Parameter()]
+        [System.String]
+        $Description,
+
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $DisplayName,
+
+        [Parameter()]
+        [System.Boolean]
+        $SupportsScopeTags,
+
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $Id,
+
+        [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance[]]
+        $Assignments,
+        #endregion
         [Parameter()]
         [System.String]
         [ValidateSet('Absent', 'Present')]
@@ -155,9 +428,10 @@ function Set-TargetResource
 
     if ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
     {
-        Write-Verbose -Message "Creating an <ResourceDescription> with <FilterKey> {$DisplayName}"
-<#AssignmentsRemove#>
-        $CreateParameters = ([Hashtable]$BoundParameters).clone()
+        Write-Verbose -Message "Creating an Intune Device Configuration Windows Team Policy for Windows10 with DisplayName {$DisplayName}"
+        $PSBoundParameters.Remove("Assignments") | Out-Null
+
+        $CreateParameters = ([Hashtable]$PSBoundParameters).clone()
         $CreateParameters = Rename-M365DSCCimInstanceParameter -Properties $CreateParameters
         $CreateParameters.Remove('Id') | Out-Null
 
@@ -168,16 +442,30 @@ function Set-TargetResource
             {
                 $CreateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $CreateParameters.$key
             }
-        }<#ResourceGenerator
+        }
         #region resource generator code
-<NewDataType>        $policy = <NewCmdLetName> <#NewKeyIdentifier#>
-<#AssignmentsNew#>        #endregionResourceGenerator#>
+        $CreateParameters.Add("@odata.type", "#microsoft.graph.windows10TeamGeneralConfiguration")
+        $policy = New-MgDeviceManagementDeviceConfiguration -BodyParameter $CreateParameters
+        $assignmentsHash = @()
+        foreach ($assignment in $Assignments)
+        {
+            $assignmentsHash += Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+        }
+
+        if ($policy.id)
+        {
+            Update-DeviceConfigurationPolicyAssignment -DeviceConfigurationPolicyId  $policy.id `
+                -Targets $assignmentsHash `
+                -Repository 'deviceManagement/deviceConfigurations'
+        }
+        #endregion
     }
     elseif ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Updating the <ResourceDescription> with <PrimaryKey> {$($currentInstance.<PrimaryKey>)}"
-<#AssignmentsRemove#>
-        $UpdateParameters = ([Hashtable]$BoundParameters).clone()
+        Write-Verbose -Message "Updating the Intune Device Configuration Windows Team Policy for Windows10 with Id {$($currentInstance.Id)}"
+        $PSBoundParameters.Remove("Assignments") | Out-Null
+
+        $UpdateParameters = ([Hashtable]$PSBoundParameters).clone()
         $UpdateParameters = Rename-M365DSCCimInstanceParameter -Properties $UpdateParameters
 
         $UpdateParameters.Remove('Id') | Out-Null
@@ -189,17 +477,29 @@ function Set-TargetResource
             {
                 $UpdateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $UpdateParameters.$key
             }
-        }<#ResourceGenerator
+        }
         #region resource generator code
-<UpdateDataType>        <UpdateCmdLetName> <#UpdateKeyIdentifier#>
-<#AssignmentsUpdate#>        #endregionResourceGenerator#>
+        $UpdateParameters.Add("@odata.type", "#microsoft.graph.windows10TeamGeneralConfiguration")
+        Update-MgDeviceManagementDeviceConfiguration  `
+            -DeviceConfigurationId $currentInstance.Id `
+            -BodyParameter $UpdateParameters
+        $assignmentsHash = @()
+        foreach ($assignment in $Assignments)
+        {
+            $assignmentsHash += Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+        }
+        Update-DeviceConfigurationPolicyAssignment `
+            -DeviceConfigurationPolicyId $currentInstance.id `
+            -Targets $assignmentsHash `
+            -Repository 'deviceManagement/deviceConfigurations'
+        #endregion
     }
     elseif ($Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Removing the <ResourceDescription> with <PrimaryKey> {$($currentInstance.<PrimaryKey>)}" <#ResourceGenerator
+        Write-Verbose -Message "Removing the Intune Device Configuration Windows Team Policy for Windows10 with Id {$($currentInstance.Id)}" 
         #region resource generator code
-<RemoveCmdLetName> <#removeKeyIdentifier#>
-        #endregionResourceGenerator#>
+Remove-MgDeviceManagementDeviceConfiguration -DeviceConfigurationId $currentInstance.Id
+        #endregion
     }
 }
 
@@ -210,7 +510,108 @@ function Test-TargetResource
     param
     (
         #region resource generator code
-<ParameterBlock><AssignmentsParam>        #endregion
+        [Parameter()]
+        [System.Boolean]
+        $AzureOperationalInsightsBlockTelemetry,
+
+        [Parameter()]
+        [System.String]
+        $AzureOperationalInsightsWorkspaceId,
+
+        [Parameter()]
+        [System.String]
+        $AzureOperationalInsightsWorkspaceKey,
+
+        [Parameter()]
+        [System.Boolean]
+        $ConnectAppBlockAutoLaunch,
+
+        [Parameter()]
+        [System.Boolean]
+        $MaintenanceWindowBlocked,
+
+        [Parameter()]
+        [System.Int32]
+        $MaintenanceWindowDurationInHours,
+
+        [Parameter()]
+        [System.TimeSpan]
+        $MaintenanceWindowStartTime,
+
+        [Parameter()]
+        [System.Boolean]
+        $MiracastBlocked,
+
+        [Parameter()]
+        [ValidateSet('userDefined','one','two','three','four','five','six','seven','eight','nine','ten','eleven','thirtySix','forty','fortyFour','fortyEight','oneHundredFortyNine','oneHundredFiftyThree','oneHundredFiftySeven','oneHundredSixtyOne','oneHundredSixtyFive')]
+        [System.String]
+        $MiracastChannel,
+
+        [Parameter()]
+        [System.Boolean]
+        $MiracastRequirePin,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockMyMeetingsAndFiles,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockSessionResume,
+
+        [Parameter()]
+        [System.Boolean]
+        $SettingsBlockSigninSuggestions,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsDefaultVolume,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsScreenTimeoutInMinutes,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsSessionTimeoutInMinutes,
+
+        [Parameter()]
+        [System.Int32]
+        $SettingsSleepTimeoutInMinutes,
+
+        [Parameter()]
+        [System.String]
+        $WelcomeScreenBackgroundImageUrl,
+
+        [Parameter()]
+        [System.Boolean]
+        $WelcomeScreenBlockAutomaticWakeUp,
+
+        [Parameter()]
+        [ValidateSet('userDefined','showOrganizerAndTimeOnly','showOrganizerAndTimeAndSubject')]
+        [System.String]
+        $WelcomeScreenMeetingInformation,
+
+        [Parameter()]
+        [System.String]
+        $Description,
+
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $DisplayName,
+
+        [Parameter()]
+        [System.Boolean]
+        $SupportsScopeTags,
+
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $Id,
+
+        [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance[]]
+        $Assignments,
+        #endregion
 
         [Parameter()]
         [System.String]
@@ -254,7 +655,7 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
-    Write-Verbose -Message "Testing configuration of the <ResourceDescription> with <PrimaryKey> {$<PrimaryKey>} and <FilterKey> {$<FilterKey>}"
+    Write-Verbose -Message "Testing configuration of the Intune Device Configuration Windows Team Policy for Windows10 with Id {$Id} and DisplayName {$DisplayName}"
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
     $ValuesToCheck = ([Hashtable]$PSBoundParameters).clone()
@@ -342,9 +743,9 @@ function Export-TargetResource
         $ManagedIdentity
     )
 
-    $ConnectionMode = New-M365DSCConnection -Workload '<#Workload#>' `
+    $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
         -InboundParameters $PSBoundParameters `
-        -ProfileName '<#APIVersion#>'
+        -ProfileName 'beta'
 
     #Ensure the proper dependencies are installed in the current environment.
     Confirm-M365DSCDependencies
@@ -359,9 +760,15 @@ function Export-TargetResource
     #endregion
 
     try
-    {<#ResourceGenerator
+    {
         #region resource generator code
-<exportGetCommand>        #endregionResourceGenerator#>
+        [array]$getValue = Get-MgDeviceManagementDeviceConfiguration `
+            -All `
+            -ErrorAction Stop | Where-Object `
+            -FilterScript { `
+                $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.windows10TeamGeneralConfiguration' `
+            }
+        #endregion
 
         $i = 1
         $dscContent = ''
@@ -375,14 +782,15 @@ function Export-TargetResource
         }
         foreach ($config in $getValue)
         {
-            $displayedKey = $config.<PrimaryKey>
+            $displayedKey = $config.Id
             if (-not [String]::IsNullOrEmpty($config.displayName))
             {
                 $displayedKey = $config.displayName
             }
             Write-Host "    |---[$i/$($getValue.Count)] $displayedKey" -NoNewline
             $params = @{
-                <PrimaryKey> = $config.<PrimaryKey><RequiredKey>
+                Id = $config.Id
+                DisplayName           =  $config.DisplayName
                 Ensure = 'Present'
                 Credential = $Credential
                 ApplicationId = $ApplicationId
@@ -395,13 +803,28 @@ function Export-TargetResource
             $Results = Get-TargetResource @Params
             $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
                 -Results $Results
-<#ConvertComplexToString#><#AssignmentsConvertComplexToString#>
+            if ($Results.Assignments)
+            {
+                $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString -ComplexObject $Results.Assignments -CIMInstanceName DeviceManagementConfigurationPolicyAssignments
+                if ($complexTypeStringResult)
+                {
+                    $Results.Assignments = $complexTypeStringResult
+                }
+                else
+                {
+                    $Results.Remove('Assignments') | Out-Null
+                }
+            }
             $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
                 -ConnectionMode $ConnectionMode `
                 -ModulePath $PSScriptRoot `
                 -Results $Results `
                 -Credential $Credential
-<#ConvertComplexToVariable#><#AssignmentsConvertComplexToVariable#><#TrailingCharRemoval#>
+            if ($Results.Assignments)
+            {
+                $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName "Assignments" -isCIMArray:$true
+            }
+
             $dscContent += $currentDSCBlock
             Save-M365DSCPartialExport -Content $currentDSCBlock `
                 -FileName $Global:PartialExportFileName
@@ -424,7 +847,71 @@ function Export-TargetResource
     }
 }
 
-<#AssignmentsFunctions#>function Rename-M365DSCCimInstanceParameter
+function Update-DeviceConfigurationPolicyAssignment
+{
+    [CmdletBinding()]
+    [OutputType([System.Collections.Hashtable])]
+    param (
+        [Parameter(Mandatory = 'true')]
+        [System.String]
+        $DeviceConfigurationPolicyId,
+
+        [Parameter()]
+        [Array]
+        $Targets,
+
+        [Parameter()]
+        [System.String]
+        $Repository = 'deviceManagement/configurationPolicies',
+
+        [Parameter()]
+        [ValidateSet('v1.0','beta')]
+        [System.String]
+        $APIVersion = 'beta'
+    )
+    try
+    {
+        $deviceManagementPolicyAssignments = @()
+        $Uri = "https://graph.microsoft.com/$APIVersion/$Repository/$DeviceConfigurationPolicyId/assign"
+
+        foreach ($target in $targets)
+        {
+            $formattedTarget = @{"@odata.type" = $target.dataType}
+            if ($target.groupId)
+            {
+                $formattedTarget.Add('groupId',$target.groupId)
+            }
+            if ($target.collectionId)
+            {
+                $formattedTarget.Add('collectionId',$target.collectionId)
+            }
+            if ($target.deviceAndAppManagementAssignmentFilterType)
+            {
+                $formattedTarget.Add('deviceAndAppManagementAssignmentFilterType',$target.deviceAndAppManagementAssignmentFilterType)
+            }
+            if ($target.deviceAndAppManagementAssignmentFilterId)
+            {
+                $formattedTarget.Add('deviceAndAppManagementAssignmentFilterId',$target.deviceAndAppManagementAssignmentFilterId)
+            }
+            $deviceManagementPolicyAssignments += @{'target' = $formattedTarget}
+        }
+        $body = @{'assignments' = $deviceManagementPolicyAssignments} | ConvertTo-Json -Depth 20
+        #write-verbose -Message $body
+        Invoke-MgGraphRequest -Method POST -Uri $Uri -Body $body -ErrorAction Stop
+    }
+    catch
+    {
+        New-M365DSCLogEntry -Message 'Error updating data:' `
+            -Exception $_ `
+            -Source $($MyInvocation.MyCommand.Source) `
+            -TenantId $TenantId `
+            -Credential $Credential
+
+        return $null
+    }
+}
+
+function Rename-M365DSCCimInstanceParameter
 {
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable], [System.Collections.Hashtable[]])]